Cambio responsabilidad de renderizado de boton y pdf de impresion de boleta de details al handler de boletas
